refactor: 체스 기물 점수를 구하는 기능을 클래스로 분할
